🐛 Rename du dossier navbar
