+++ frontend/app/page.tsx
@@ -1,4 +1,4 @@
-import AppSidebar from "@/components/navab/Sidebar";
+import AppSidebar from "@/components/navbar/Sidebar";
 
 export default function Home() {
   return <div>Dashboard</div>;
