+++ frontend/components/navbar/Sidebar.tsx
